refactor(Worker): Rename UAPI Plugin to Model
