+++ go/worker/plugins/model/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "plugin.go",
         "starlark_module.go",
     ],
-    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/plugins/uapi",
+    importpath = "github.com/michelangelo-ai/michelangelo/go/worker/plugins/model",
     visibility = ["//visibility:public"],
     deps = [
         "//go/worker/activities/model:go_default_library",
