+++ go/worker/plugins/model/starlark_module.go
@@ -1,4 +1,4 @@
-package uapi
+package model
 
 import (
 	"fmt"
