+++ go/worker/starlark/BUILD.bazel
@@ -7,10 +7,10 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//go/worker/plugins/cachedoutput:go_default_library",
+        "//go/worker/plugins/model:go_default_library",
         "//go/worker/plugins/ray:go_default_library",
         "//go/worker/plugins/spark:go_default_library",
         "//go/worker/plugins/storage:go_default_library",
-        "//go/worker/plugins/uapi:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//service:go_default_library",
         "@com_github_cadence_workflow_starlark_worker//worker:go_default_library",
         "@org_uber_go_fx//:go_default_library",
