+++ go/worker/starlark/module.go
@@ -8,10 +8,10 @@ import (
 	"go.uber.org/fx"
 
 	"github.com/michelangelo-ai/michelangelo/go/worker/plugins/cachedoutput"
+	"github.com/michelangelo-ai/michelangelo/go/worker/plugins/model"
 	"github.com/michelangelo-ai/michelangelo/go/worker/plugins/ray"
 	"github.com/michelangelo-ai/michelangelo/go/worker/plugins/spark"
 	"github.com/michelangelo-ai/michelangelo/go/worker/plugins/storage"
-	"github.com/michelangelo-ai/michelangelo/go/worker/plugins/uapi"
 )
 
 // RegisterStoragePlugin adds the storage plugin to the plugin registry.
@@ -34,9 +34,9 @@ func RegisterSparkPlugin(registry map[string]service.IPlugin) {
 	registry[spark.Plugin.ID()] = spark.Plugin
 }
 
-// RegisterUAPIPlugin adds the uapi plugin to the plugin registry.
-func RegisterUAPIPlugin(registry map[string]service.IPlugin) {
-	registry[uapi.Plugin.ID()] = uapi.Plugin
+// RegisterModelPlugin adds the model plugin to the plugin registry.
+func RegisterModelPlugin(registry map[string]service.IPlugin) {
+	registry[model.Plugin.ID()] = model.Plugin
 }
 
 // CreateStarlarkService creates the starlark service with all registered plugins.
@@ -61,6 +61,6 @@ var Module = fx.Options(
 	fx.Invoke(RegisterCachedOutputPlugin),
 	fx.Invoke(RegisterRayPlugin),
 	fx.Invoke(RegisterSparkPlugin),
-	fx.Invoke(RegisterUAPIPlugin),
+	fx.Invoke(RegisterModelPlugin),
 	fx.Invoke(CreateStarlarkService),
 )
